toolkit/components/windowwatcher/test/test_modal_windows.html

<!DOCTYPE HTML>
<html>
<!--
Test that the parent can open modal windows, and that the modal window
that is opened reports itself as being modal.
-->
<head>
  <meta charset="utf-8">
  <title>Test a modal window</title>

  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">

  <script type="application/javascript">

  const {BrowserTestUtils} = ChromeUtils.import("resource://testing-common/BrowserTestUtils.jsm");

  add_task(async function() {
    BrowserTestUtils.domWindowOpened().then((win) => {
      let treeOwner = win.docShell.treeOwner;
      let chromeFlags = treeOwner.QueryInterface(Ci.nsIInterfaceRequestor)
                                 .getInterface(Ci.nsIAppWindow)
                                 .chromeFlags;
      ok(chromeFlags & Ci.nsIWebBrowserChrome.CHROME_MODAL,
         "Should have the modal chrome flag");

      let wbc = treeOwner.QueryInterface(Ci.nsIInterfaceRequestor)
                         .getInterface(Ci.nsIWebBrowserChrome);
      ok(wbc.isWindowModal(), "Should report as modal");

      win.close();
    });

    window.openDialog("data:text/html,<p>This is a modal window for test_modal_windows.html</p>",
                      "_blank", "modal,noopener", null);
    // Since the modal runs a nested event loop, just to be on the safe side,
    // we'll wait a tick of the main event loop before resolving the task.
    await new Promise(resolve => setTimeout(resolve, 0));
  });

  </script>
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none">
</div>
<pre id="test">
</pre>
</body>
</html>
